stock_picker_agent/requirements.txt
crewai
crewai-tools
langchain-openai
yfinance
pandas
numpy
python-dotenv
requests
beautifulsoup4